+++ frontend/pubspec.yaml
@@ -42,6 +42,7 @@ dependencies:
   google_sign_in: ^7.1.1
   google_sign_in_android: ^7.0.3
   shared_preferences: ^2.5.3
+  path_provider: ^2.1.5
 
 dev_dependencies:
   flutter_test:
